Allow `source` to be a URL-returning function.
Currently, when `source` is string-valued it naively appends the typed
input ("display" data). This limits it's usefulness against more
complicated endpoints.

Permit `source` to be a function which receives the typed input and must
return a string. The string will be used as an endpoint URL.

(A future improvement would allow the function to return a `Request`
object, to better expose the complete Fetch API.)